pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docu-processing"
version = "0.1.0"
description = "Document processing and RAG system using Docling, LangChain, and ChromaDB"
readme = "readme.md"
requires-python = ">=3.11,<3.12"
authors = [
    {name = "Arindam Samanta", email = "samantoz@gmail.com"}
]
license = {text = "MIT"}
keywords = ["document-processing", "rag", "embeddings", "pdf", "langchain", "chromadb"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "langchain>=0.3.26",
    "langchain-openai>=0.3.28",
    "langchain-community>=0.0.20",
    "openai>=1.97.0",
    "chromadb>=1.0.15",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "tiktoken>=0.5.0",
    "python-dotenv>=1.1.1",
    "filetype>=1.2.0",
    "pypdf>=3.0.0",
    "pymupdf>=1.23.0",
    "docling>=2.61.2",
    "ollama>=0.6.0",
    "ollama-ocr>=0.1.6",
    "rapidocr>=3.4.2",
    "easyocr>=1.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
all = [
    "docu-processing[dev]",
]

[project.urls]
Homepage = "https://github.com/samantoz/docu-processing"
Repository = "https://github.com/samantoz/docu-processing"
Issues = "https://github.com/samantoz/docu-processing/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["scripts*"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Note: docling and torch dependencies removed from main dependencies due to platform compatibility issues
# For macOS x86_64, install separately:
# uv pip install docling --index-url https://download.pytorch.org/whl/cpu --extra-index-url https://pypi.org/simple
# 
# For other platforms, install normally:
# uv pip install docling torch torchvision
